+++ andes/routines/tds.py
@@ -782,15 +782,35 @@ class TDS(BaseRoutine):
 
     def fg_update(self, models):
         """
-        Update `f` and `g` equations.
+        Perform one round of evaluation for one iteration step.
+        The following operations are performed in order:
+
+        - discrete flags updating through ``l_update_var``
+        - variable service updating through ``s_update_var``
+        - evaluation of the right-hand-side of ``f``
+        - equation-dependent discrete flags updating through ``l_update_eq``
+        - evaluation of the right-hand-side of ``g``
+        - collection of residuals into dae through ``fg_to_dae``.
+
         """
+
         system = self.system
         system.dae.clear_fg()
-        system.l_update_var(models=models, niter=self.niter, err=self.mis[-1])
+
+        system.l_update_var(models=models,
+                            niter=self.niter,
+                            err=self.mis[-1],
+                            )
+
         system.s_update_var(models=models)  # update VarService
+
+        # evalute the RHS of `f` and check the limiters (anti-windup)
+        # 12/08/2020: Moded `l_update_eq` to before `g_update`
+        #   because some algebraic variables depend on pegged states.
         system.f_update(models=models)
-        system.g_update(models=models)
         system.l_update_eq(models=models)
+
+        system.g_update(models=models)
         system.fg_to_dae()
 
     def _fg_wrapper(self, xy):